Bookmark refactor
