+++ src/components/bookmark_icon.js
@@ -1,0 +1,13 @@
+import React from "react"
+
+import Icon from "./icon"
+
+const BookmarkIcon = props => (
+  <Icon
+    {...props}
+    viewBox="0 0 26 34"
+    pathD="M2 0h22a2 2 0 012 2v29.586a1 1 0 01-1.707.707L13.707 21.707a1 1 0 00-1.414 0L1.707 32.293A1 1 0 010 31.586V2a2 2 0 012-2z"
+  />
+)
+
+export default BookmarkIcon
