+++ src/components/icon.css
@@ -1,0 +1,3 @@
+svg.Icon {
+  overflow: visible;
+}
