+++ src/components/project.css
@@ -157,3 +157,26 @@
   overflow: hidden;
   text-overflow: ellipsis;
 }
+
+.Project--bookmark {
+  visibility: hidden;
+  position: absolute;
+  top: -2px;
+  right: 26px;
+  width: 26px;
+  height: 34px;
+  border-radius: 2px 2px 5px 5px;
+  z-index: 1;
+}
+
+.Project--bookmark[data-is-bookmarked="true"] {
+  visibility: visible;
+}
+
+.Project--bookmark > svg {
+  overflow: visible;
+}
+
+.Project--bookmark:hover {
+  opacity: 0.75;
+}
