+++ src/templates/project.js
@@ -2,11 +2,12 @@ import React from "react"
 import { graphql, Link } from "gatsby"
 import Img from "gatsby-image"
 
-import Bookmark from "../components/bookmark"
+import BookmarkIcon from "../components/bookmark_icon"
 import Layout from "../components/layout"
 import Markdown from "../components/markdown"
 import RelatedProject from "../components/project"
 import SEO from "../components/seo"
+import useBookmarkState from "../components/bookmark_state"
 
 import { flatten, normalizeCollection } from "../utils"
 
@@ -65,6 +66,8 @@ const Project = ({
     projects.find(({ id }) => id === project.id)
   )
 
+  const { bookmarked, loaded, toggleBookmark } = useBookmarkState(project.slug)
+
   return (
     <Layout>
       <SEO title={project.name} />
@@ -100,7 +103,23 @@ const Project = ({
                 ))}
               </div>
 
-              <Bookmark project={project} />
+              <div
+                className="ProjectPage--header-action-bookmark"
+                data-is-bookmarked={bookmarked}
+                data-is-loaded={loaded}
+              >
+                <button
+                  className="ProjectPage--header-action-button Button"
+                  onClick={toggleBookmark}
+                >
+                  <span className="ProjectPage--header-action-bookmark-icon">
+                    <BookmarkIcon withGradientFill={bookmarked} />
+                  </span>
+                  <span className="ProjectPage--header-action-bookmark-text">
+                    {bookmarked ? "Bookmarked" : "Bookmark"}
+                  </span>
+                </button>
+              </div>
             </div>
           )}
         </div>
